Suppress metro warnings
