--- src/polyfillTextureLoader.fx.web.ts
@@ -1,1 +1,0 @@
-// noop
