--- src/suppressWarnings.web.ts
@@ -1,1 +1,0 @@
-export default (shouldSuppress: boolean = true) => {};
